Split benches into two files
