+++ node-graph/graph-craft/src/lib.rs
@@ -12,3 +12,6 @@ pub mod graphene_compiler;
 pub mod imaginate_input;
 
 pub mod wasm_application_io;
+
+#[cfg(feature = "loading")]
+pub mod util;
